third_party/WebKit/LayoutTests/fast/forms/select/ignore-keydown-with-modifiers.html

<!DOCTYPE html>
<body>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<div id="log"></div>

<select autofocus>
<option selected>a</option>
<option>b</option>
<option>c</option>
</select>

<script>

test(function() {
    var select = document.querySelector('select');
    assert_not_equals(null, select);

    select.focus();
    assert_equals(select, document.activeElement, 'has focus');

    assert_equals(select.selectedIndex, 0, 'starts at 0');

    eventSender.keyDown('ArrowRight', ['shiftKey']);
    assert_equals(select.selectedIndex, 0, 'shift+right is ignored');

    eventSender.keyDown('ArrowRight', ['ctrlKey']);
    assert_equals(select.selectedIndex, 0, 'ctrl+right is ignored');

    eventSender.keyDown('ArrowRight', ['altKey']);
    assert_equals(select.selectedIndex, 0, 'alt+right is ignored');

    eventSender.keyDown('ArrowRight', ['metaKey']);
    assert_equals(select.selectedIndex, 0, 'meta+right is ignored');
}, 'Tests that arrow keys with modifiers are ignored');

</script>
</body>
